Refactor tree filtering to be dynamic.
ApplyFilterToTreeTableAction didn't really apply a filter.
It just wiped out all the nodes in the tree, and added the root resource nodes back in.
From then on, loading new nodes was filtered.  This makes filtering quite a destructive action
while exploring a tree, as it wipes out all the paths to the places you were interested in before you set the filter.

These changes remove ApplyFilterToTreeTableAction with a call to ProfileForm.refreshFiltering.

A new method wired through the profile instance managers obtains the filter status for any set of
parents ids and a given filter.  So instead of wiping out all the nodes in the tree, we just get and set an updated filter status for all the nodes, and tell the TreeTableModel to filter out the ones
which have status 0.

You can now flip between filtered and unfiltered dynamically, preserving all the folder expansions
you've built up exploring the tree.